+++ sphinx/application/common/wrappers/wrapper-common/src/main/java/chat/sphinx/wrapper_common/lightning/LightningNodeDescriptor.kt
@@ -1,0 +1,9 @@
+package chat.sphinx.wrapper_common.lightning
+
+/**
+ * A class that contains data which is descriptive of how
+ * to reach one's lightning node.
+ * */
+sealed interface LightningNodeDescriptor {
+    val value: String
+}