+++ sphinx/application/data/features/feature-repository/src/main/java/chat/sphinx/feature_repository/util/Extensions.kt
@@ -324,6 +324,24 @@ fun TransactionCallbacks.upsertMessage(dto: MessageDto, queries: SphinxDatabaseQ
         ChatId(it)
     } ?: ChatId(ChatId.NULL_CHAT_ID.toLong())
 
+    dto.media_token?.let { mediaToken ->
+        dto.media_type?.let { mediaType ->
+
+            if (mediaToken.isEmpty() || mediaType.isEmpty()) return
+
+            queries.messageMediaUpsert(
+                dto.media_key?.toMediaKey(),
+                mediaType.toMediaType(),
+                MediaToken(mediaToken),
+                MessageId(dto.id),
+                chatId,
+                dto.mediaKeyDecrypted?.toMediaKeyDecrypted(),
+                dto.mediaLocalFile,
+            )
+
+        }
+    }
+
     queries.messageUpsert(
         dto.status.toMessageStatus(),
         dto.seenActual.toSeen(),
@@ -358,9 +376,8 @@ fun TransactionCallbacks.upsertMessage(dto: MessageDto, queries: SphinxDatabaseQ
                 MessageId(dto.id),
                 chatId,
                 dto.mediaKeyDecrypted?.toMediaKeyDecrypted(),
-                dto.mediaLocalFile,
+                dto.mediaLocalFile
             )
-
         }
     }
 }
@@ -394,4 +411,13 @@ inline fun TransactionCallbacks.deleteContactById(
     queries.contactDeleteById(contactId)
     queries.inviteDeleteByContactId(contactId)
     queries.dashboardDeleteById(contactId)
+}
+
+@Suppress("NOTHING_TO_INLINE")
+inline fun TransactionCallbacks.deleteMessageById(
+    messageId: MessageId,
+    queries: SphinxDatabaseQueries
+) {
+    queries.messageDeleteById(messageId)
+    queries.messageMediaDeleteById(messageId)
 }