+++ sphinx/application/network/concepts/concept-link-preview/src/main/java/chat/sphinx/concept_link_preview/model/HtmlPreviewFavIconUrl.kt
@@ -1,0 +1,20 @@
+package chat.sphinx.concept_link_preview.model
+
+import okhttp3.HttpUrl.Companion.toHttpUrlOrNull
+
+@Suppress("NOTHING_TO_INLINE")
+inline fun String.toHtmlPreviewFavIconUrlOrNull(): HtmlPreviewFavIconUrl? =
+    try {
+        HtmlPreviewFavIconUrl(this)
+    } catch (e: IllegalArgumentException) {
+        null
+    }
+
+@JvmInline
+value class HtmlPreviewFavIconUrl(val value: String) {
+    init {
+        require(value.toHttpUrlOrNull() != null) {
+            "HtmlPreviewFavIconUrl was not a valid url"
+        }
+    }
+}