+++ sphinx/screens/chats/chat-contact/chat-contact/src/main/java/chat/sphinx/chat_contact/ui/ChatContactViewModel.kt
@@ -32,6 +32,7 @@ import chat.sphinx.wrapper_message.Message
 import dagger.hilt.android.lifecycle.HiltViewModel
 import io.matthewnelson.android_feature_navigation.util.navArgs
 import io.matthewnelson.concept_coroutines.CoroutineDispatchers
+import chat.sphinx.concept_link_preview.LinkPreviewHandler
 import io.matthewnelson.concept_media_cache.MediaCacheHandler
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.*
@@ -62,6 +63,7 @@ internal class ChatContactViewModel @Inject constructor(
     mediaCacheHandler: MediaCacheHandler,
     savedStateHandle: SavedStateHandle,
     cameraViewModelCoordinator: ViewModelCoordinator<CameraRequest, CameraResponse>,
+    linkPreviewHandler: LinkPreviewHandler,
     LOG: SphinxLogger,
 ): ChatViewModel<ChatContactFragmentArgs>(
     app,
@@ -75,6 +77,7 @@ internal class ChatContactViewModel @Inject constructor(
     mediaCacheHandler,
     savedStateHandle,
     cameraViewModelCoordinator,
+    linkPreviewHandler,
     LOG,
 ) {
     override val args: ChatContactFragmentArgs by savedStateHandle.navArgs()